+++ packages/gallery/tests/drivers/mocks/styles.js
@@ -29,6 +29,7 @@ const styleParams = {
   imageMargin: 10,
   galleryMargin: 0,
   scatter: 0,
+  rotatingScatter: '',
   gridStyle: 0,
   mobilePanorama: false,
   placeGroupsLtr: false,