+++ packages/lib/src/common/defaultStyles.js
@@ -28,6 +28,7 @@ const defaultStyles = {
   imageMargin: 10,
   galleryMargin: 0,
   scatter: 0,
+  rotatingScatter: '',
   gridStyle: 0,
   mobilePanorama: false,
   placeGroupsLtr: false,
@@ -141,6 +142,7 @@ const defaultStyles = {
   videoSound: false,
   videoSpeed: '1',
   videoLoop: true,
+  jsonStyleParams: '',
 };
 
 /* Object.entries(galleryOptions).forEach(([styleParam, settings]) => {