+++ CHANGELOG.md
@@ -1,3 +1,61 @@
+## [v2.4.15](https://pro-gallery-2-4-15.surge.sh) (25/01/2021)
+ 
+#### GALLERY
+ -  improve: pass gallery item on CURRENT_ITEM_CHANGED event 
+
+---
+## [v2.4.14](https://pro-gallery-2-4-14.surge.sh) (24/01/2021)
+ 
+#### GALLERY
+ -  fix: use galleryWidth when calculating cssScroll for horizontal animations
+ -  fix: wrap console.log
+
+#### OTHER
+ - Update README.md
+
+---
+## [v2.4.13](https://pro-gallery-2-4-13.surge.sh) (17/01/2021)
+ 
+#### GALLERY
+ -  (galleryContainerNew): in debouncedReCreateGallery, call reCreateGalleryExpensively with the correct params.
+
+---
+## [v2.4.12](https://pro-gallery-2-4-12.surge.sh) (14/01/2021)
+ 
+#### GALLERY
+ -  fix blueprints: height change between layouts
+
+---
+## [v2.4.11](https://pro-gallery-2-4-11.surge.sh) (14/01/2021)
+ 
+#### GALLERY
+ -  layoutingIndex fix: allow looping on items using the blueprintsManager
+ -  feature (overlayBackgroundGradient): add the option to set the gradient of the overlay
+
+---
+## [v2.4.10](https://pro-gallery-2-4-10.surge.sh) (13/01/2021)
+ 
+#### GALLERY
+ -  Presets: Handle designed presets with new 'jsonStyleParams'
+
+---
+## [v2.4.9](https://pro-gallery-2-4-9.surge.sh) (07/01/2021)
+ 
+#### GALLERY
+ -  fix: black stripes in videos in grid fill
+
+---
+## [v2.4.8](https://pro-gallery-2-4-8.surge.sh) (07/01/2021)
+ 
+#### OTHER
+ - Added new designed presets
+ - Added rotatingScatter
+ - Added alternate placement
+
+#### LAYOUTS
+ -  fix (strip): do not allow incomplete strips with fixed groups per row
+
+---
 ## [v2.4.7](https://pro-gallery-2-4-7.surge.sh) (04/01/2021)
  
 #### GALLERY